requirements.txt
telethon
openai
textblob
tinydb
python-dotenv
